+++ sample/src/commonMain/kotlin/com/github/panpf/zoomimage/sample/AppSettings.common.kt
@@ -64,6 +64,8 @@ expect class AppSettings(context: PlatformContext) {
 
     val limitOffsetWithinBaseVisibleRect: SettingsStateFlow<Boolean>
 
+    val containerWhitespaceMultiple: SettingsStateFlow<Float>
+
     val readModeEnabled: SettingsStateFlow<Boolean>
 
     val readModeAcceptedBoth: SettingsStateFlow<Boolean>