+++ sample/src/nonAndroidMain/kotlin/com/github/panpf/zoomimage/sample/AppSettings.nonAndroid.kt
@@ -14,6 +14,7 @@ import com.github.panpf.zoomimage.sample.ui.util.name
 import com.github.panpf.zoomimage.sample.util.ParamLazy
 import com.github.panpf.zoomimage.sample.util.SettingsStateFlow
 import com.github.panpf.zoomimage.sample.util.booleanSettingsStateFlow
+import com.github.panpf.zoomimage.sample.util.floatSettingsStateFlow
 import com.github.panpf.zoomimage.sample.util.intSettingsStateFlow
 import com.github.panpf.zoomimage.sample.util.isDebugMode
 import com.github.panpf.zoomimage.sample.util.stateMap
@@ -105,6 +106,10 @@ actual class AppSettings actual constructor(val context: PlatformContext) {
 
     actual val limitOffsetWithinBaseVisibleRect: SettingsStateFlow<Boolean> by lazy {
         booleanSettingsStateFlow(context, "limitOffsetWithinBaseVisibleRect", false)
+    }
+
+    actual val containerWhitespaceMultiple: SettingsStateFlow<Float> by lazy {
+        floatSettingsStateFlow(context, "containerWhitespaceMultiple1", 0f)
     }
 
     actual val readModeEnabled: SettingsStateFlow<Boolean> by lazy {